change endpoint name for blogs for certain production
